// test/fields/regressions/frogbot-types.ts
export interface FieldArgument {
  id: number;
  blocks?:
    | {
        observed?: string | null;
        validated?: string | null;
        readable?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'observed-block';
      }[]
    | null;
  payloadAuthored?: string | null;
  updatedAt: string;
  createdAt: string;
}